requirements.txt
PyYAML
pyad
hvac
openpyxl
pywin32
pandas

## pip install -r requirements.txt

